veo3-video-generator/requirements.txt
google-genai>=0.1.0
argparse
pathlib
